+++ src/game/Levels.h
@@ -53,6 +53,6 @@ long GetLevelNumByName(const std::string & name);
 const char * GetLevelNameByNum(long num);
 
 //! Get the physical level (depth) of a "subdivided" level
-long ARX_LEVELS_GetRealNum(long num);
+int ARX_LEVELS_GetRealNum(long num);
 
 #endif // ARX_GAME_LEVELS_H
